+++ examples/terraform/main.tf
@@ -1,117 +1,85 @@
-terraform {
-  required_providers {
-    pagerduty = {
-      source  = "PagerDuty/pagerduty"
-      version = "1.9.6"
+module "pagerduty-service" {
+  source = "git@github.com:adamwshero/terraform-pagerduty-service.git//.?ref=1.1.0"
+
+  // PagerDuty Service
+  name              = "My Critical Service"
+  description       = "Service for all prod services."
+  escalation_policy = "My Escalation Policy Name"
+  alert_creation    = "create_alerts_and_incidents"
+  resolve_timeout   = 14400
+  ack_timeout       = 600
+  token             = file("${path.module}/my_pagerduty_api_key.yaml")
+
+  // Maintenance Windows
+  enable_maintenance_windows = true
+  maintenance_windows = [
+    {
+      description = "Overnight Maintenance"
+      start_time  = "2022-11-09T20:00:00-05:00"
+      end_time    = "2022-11-09T22:00:00-05:00"
+    },
+    {
+      description = "Weekend Maintenance"
+      start_time  = "2022-12-09T20:00:00-05:00"
+      end_time    = "2022-12-09T22:00:00-05:00"
     }
-  }
-}
-provider "pagerduty" {
-  token = var.token
-}
+  ]
 
-#######################
-##  PagerDuty Service
-#######################
+   // Incident Urgency Rules
+  incident_urgency_rule = [{
+    type    = "constant"
+    urgency = "low"
 
-data "pagerduty_escalation_policy" "this" {
-  name = var.escalation_policy
-}
+    during_support_hours = [{
+      type    = "constant"
+      urgency = "high"
+    }]
+    outside_support_hours = [{
+      type    = "constant"
+      urgency = "low"
+    }]
+  }]
 
-resource "pagerduty_service" "this" {
-  name                    = var.name
-  auto_resolve_timeout    = var.resolve_timeout
-  acknowledgement_timeout = var.ack_timeout
-  escalation_policy       = data.pagerduty_escalation_policy.this.id
-  alert_creation          = var.alert_creation
-}
-
-######################################
-##  PagerDuty CloudWatch Integration
-######################################
-
-data "pagerduty_vendor" "this" {
-  name = "CloudWatch"
-}
-
-resource "pagerduty_service_integration" "this" {
-  name    = data.pagerduty_vendor.this.name
-  service = pagerduty_service.this.id
-  vendor  = data.pagerduty_vendor.this.id
-  # type    = (do not use for Datadog or Cloudwatch "vendor" integrations..only for generic service integrations)
-}
-
-#############################
-##  SNS Topic For PagerDuty
-#############################
-
-resource "aws_sns_topic" "this" {
-  name = "${var.prefix}-${var.service_name}"
-}
-
-######################################
-##  SNS/PagerDuty Topic Subscription
-######################################
-
-resource "aws_sns_topic_subscription" "this" {
-  topic_arn              = aws_sns_topic.this.arn
-  protocol               = "https"
-  endpoint_auto_confirms = true #required or the subscription won't auto-confirm
-  endpoint               = "https://events.pagerduty.com/integration/${pagerduty_service_integration.this.integration_key}/enqueue"
-
-}
-
-###############################
-##  PagerDuty Slack Extension
-###############################
-
-data "pagerduty_extension_schema" "this" {
-  name = var.schema_webhook
-}
-
-resource "pagerduty_extension" "this" {
-  name              = "DevOps: Slack"
-  endpoint_url      = var.url
-  extension_schema  = data.pagerduty_extension_schema.this.id
-  extension_objects = [pagerduty_service.this.id]
-
-  config = <<EOF
-  {
-      "app_id": "${var.app_id}",
-      "authed_user": {
-        "id": "${var.authed_user}"
-      },
-      "bot_user_id": "${var.bot_user_id}",
-      "enterprise": "null",
-      "incoming_webhook": {
-        "channel": "${var.slack_channel}",
-        "channel_id": "${var.slack_channel_id}",
-        "configuration_url": "${var.configuration_url}",
-        "url": "${var.url}"
-      },
-      "restrict": "any",
-      "is_enterprise_install": "false",
-      "notify_types": {
-        "resolve": "${var.notify_resolve}",
-        "trigger": "${var.notify_trigger}",
-        "escalate": "${var.notify_escalate}",
-        "acknowledge": "${var.notify_acknowledge}",
-        "assignments": "${var.notify_assignments}",
-        "annotate": "${var.notify_annotate}"
-      },
-      "ok": true,
-      "referer": "${var.referer}",
-      "scope": "app_mentions:read,channels:join,channels:manage,channels:read,chat:write,chat:write.public,commands,groups:read,groups:write,im:read,im:write,incoming-webhook,pins:write,team:read,users:read,workflow.steps:execute",
-      "team": {
-        "id": "${var.slack_team_id}",
-        "name": "${var.slack_team_name}"
-      },
-      "token_type": "bot",
-      "urgency": {
-        "high": "${var.high_urgency}",
-        "low": "${var.low_urgency}"
-      }
+  // Support Hours
+  support_hours = [
+    {
+      type         = "fixed_time_per_day"
+      time_zone    = "America/Lima"
+      days_of_week = [1, 2, 3, 4, 5]
+      start_time   = "05:00:00"
+      end_time     = "16:00:00"
     }
-EOF
+  ]
 
+  // Service Integration
+  enable_service_integration = true
+  vendor_name                = "CloudWatch"
+
+  // SNS Topic
+  create_sns_topic = true
+  service_name     = "AcmeCorp-Elasticsearch"
+
+  // PagerDuty Extension
+  create_extension       = true
+  extension_name         = "DevOps: Slack"
+  schema_webhook         = "Generic V1 Webhook"
+  config = templatefile("${path.module}/slack/config.json.tpl", {    
+    app_id             = "A1AAAAAAA"
+    authed_user        = "A11AAA11AAA"
+    bot_user_id        = "A111AAAA11A"
+    slack_team_id      = "AAAAAA11A"
+    slack_team_name    = "AcmeCorp"
+    slack_channel      = "#devops-pagerduty"
+    slack_channel_id   = "A11AA1AAA1A"
+    configuration_url  = "https://acme-corp.slack.com/services/A111AAAAAAAA"
+    referer            = "https://acmecorp.pagerduty.com/services/A1AAAA1/integrations?service_profile=1"
+    notify_resolve     = true
+    notify_trigger     = true
+    notify_escalate    = true
+    notify_acknowledge = true
+    notify_assignments = true
+    notify_annotate    = true
+    high_urgency       = true
+    low_urgency        = true
+  })
 }