+++ main.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     pagerduty = {
       source  = "PagerDuty/pagerduty"
-      version = "2.5.0"
+      version = "2.6.2"
     }
   }
 }
@@ -10,14 +10,9 @@ provider "pagerduty" {
   token = var.token
 }
 
-#######################
-##  PagerDuty Service
-#######################
-
 data "pagerduty_escalation_policy" "this" {
   name = var.escalation_policy
 }
-
 resource "pagerduty_service" "this" {
   name                    = var.name
   description             = var.description
@@ -27,20 +22,21 @@ resource "pagerduty_service" "this" {
   alert_creation          = var.alert_creation
   dynamic "incident_urgency_rule" {
     for_each = var.incident_urgency_rule
+    iterator = incident_urgency_rule
     content {
-      type = incident_urgency_rule.value["type"]
+      type    = incident_urgency_rule.value["type"]
       urgency = incident_urgency_rule.value["urgency"]
       dynamic "during_support_hours" {
         for_each = incident_urgency_rule.value.during_support_hours
         content {
-          type = during_support_hours.value["type"]
+          type    = during_support_hours.value["type"]
           urgency = during_support_hours.value["urgency"]
         }
-      } 
+      }
       dynamic "outside_support_hours" {
         for_each = incident_urgency_rule.value.outside_support_hours
         content {
-          type = outside_support_hours.value["type"]
+          type    = outside_support_hours.value["type"]
           urgency = outside_support_hours.value["urgency"]
         }
       }
@@ -49,10 +45,10 @@ resource "pagerduty_service" "this" {
   dynamic "support_hours" {
     for_each = var.support_hours
     content {
-      type = support_hours.value["type"]
-      start_time = support_hours.value["start_time"]
-      end_time = support_hours.value["end_time"]
-      time_zone = support_hours.value["time_zone"]
+      type         = support_hours.value["type"]
+      start_time   = support_hours.value["start_time"]
+      end_time     = support_hours.value["end_time"]
+      time_zone    = support_hours.value["time_zone"]
       days_of_week = support_hours.value["days_of_week"]
     }
   }
@@ -60,7 +56,7 @@ resource "pagerduty_service" "this" {
     for_each = var.scheduled_actions
     content {
       to_urgency = scheduled_actions.value["to_urgency"]
-      type = scheduled_actions.value["type"]
+      type       = scheduled_actions.value["type"]
       dynamic "at" {
         for_each = scheduled_actions.value.at
         content {
@@ -71,93 +67,53 @@ resource "pagerduty_service" "this" {
     }
   }
 }
-
-######################################
-##  PagerDuty CloudWatch Integration
-######################################
-
 data "pagerduty_vendor" "this" {
-  name = "CloudWatch"
+  name = var.vendor_name
 }
 
 resource "pagerduty_service_integration" "this" {
-  name    = data.pagerduty_vendor.this.name
-  service = pagerduty_service.this.id
+  count = var.enable_service_integration ? 1 : 0
+
+  name    = data.pagerduty_vendor.this.name  
   vendor  = data.pagerduty_vendor.this.id
+  service = pagerduty_service.this.id
   # type    = (do not use for Datadog or Cloudwatch "vendor" integrations..only for generic service integrations)
 }
-
-#############################
-##  SNS Topic For PagerDuty
-#############################
-
 resource "aws_sns_topic" "this" {
-  name = "${var.prefix}-${var.service_name}"
+  count = var.create_sns_topic ? 1 : 0
+
+  name = var.service_name
 }
-
-######################################
-##  SNS/PagerDuty Topic Subscription
-######################################
-
 resource "aws_sns_topic_subscription" "this" {
-  topic_arn              = aws_sns_topic.this.arn
+  count = var.create_sns_topic ? 1 : 0
+
+  topic_arn              = aws_sns_topic.this[0].arn
   protocol               = "https"
   endpoint_auto_confirms = true #required or the subscription won't auto-confirm
-  endpoint               = "https://events.pagerduty.com/integration/${pagerduty_service_integration.this.integration_key}/enqueue"
+  endpoint               = "https://events.pagerduty.com/integration/${pagerduty_service_integration.this[0].integration_key}/enqueue"
 
 }
-
-###############################
-##  PagerDuty Slack Extension
-###############################
-
 data "pagerduty_extension_schema" "this" {
-  name = var.schema_webhook
+  count = var.create_extension ? 1 : 0
+
+  name = var.schema_name
+}
+resource "pagerduty_extension" "this" {
+  count = var.create_extension ? 1 : 0
+
+  name              = var.extension_name
+  endpoint_url      = var.endpoint_url
+  extension_schema  = data.pagerduty_extension_schema.this[count.index].id
+  extension_objects = [pagerduty_service.this.id]
+  config = var.config
 }
 
-resource "pagerduty_extension" "this" {
-  name              = "DevOps: Slack"
-  endpoint_url      = var.url
-  extension_schema  = data.pagerduty_extension_schema.this.id
-  extension_objects = [pagerduty_service.this.id]
-
-  config = <<EOF
-  {
-      "app_id": "${var.app_id}",
-      "authed_user": {
-        "id": "${var.authed_user}"
-      },
-      "bot_user_id": "${var.bot_user_id}",
-      "enterprise": "null",
-      "incoming_webhook": {
-        "channel": "${var.slack_channel}",
-        "channel_id": "${var.slack_channel_id}",
-        "configuration_url": "${var.configuration_url}",
-        "url": "${var.url}"
-      },
-      "restrict": "any",
-      "is_enterprise_install": "false",
-      "notify_types": {
-        "resolve": "${var.notify_resolve}",
-        "trigger": "${var.notify_trigger}",
-        "escalate": "${var.notify_escalate}",
-        "acknowledge": "${var.notify_acknowledge}",
-        "assignments": "${var.notify_assignments}",
-        "annotate": "${var.notify_annotate}"
-      },
-      "ok": true,
-      "referer": "${var.referer}",
-      "scope": "app_mentions:read,channels:join,channels:manage,channels:read,chat:write,chat:write.public,commands,groups:read,groups:write,im:read,im:write,incoming-webhook,pins:write,team:read,users:read,workflow.steps:execute",
-      "team": {
-        "id": "${var.slack_team_id}",
-        "name": "${var.slack_team_name}"
-      },
-      "token_type": "bot",
-      "urgency": {
-        "high": "${var.high_urgency}",
-        "low": "${var.low_urgency}"
-      }
-    }
-EOF
-
+resource "pagerduty_maintenance_window" "this" {
+  for_each = {
+    for k,v in var.maintenance_windows : k => v if var.enable_maintenance_windows == true
+  }
+    start_time  = each.value.start_time
+    end_time    = each.value.end_time
+    description = each.value.description
+    services    = [pagerduty_service.this.id]
 }