+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-## pagerduty-v1.0.10 (June 6, 2022)
+## 1.1.0 (September 27, 2022)
+
+CHORE:
+* Updated pinned "PagerDuty/pagerduty" provider to 2.6.2
+* Added outputs for Slack Extension
+* Added outputs for Service Integration
+* Updated all examples
+* Added examples for many scenarios
+* Expanded README
+
+ENHANCEMENT: (BREAKING CHANGES!)
+* Simplified incident_urgency and scheduled actions map variables
+* Removed Extension configuration from module and into input block (templatefile) for more control.
+* Added option to disable/enable Slack extension
+* Added option to disable/enable Service integration
+* Added option to disable/enable SNS Topic
+* Removed hard-coded "CloudWatch" vendor integration. Allowed to be dynamic input now.
+* Renamed `shema_webhook` variable to `schema_name`
+
+FEATURE:
+* Support for one or many maintenance windows.
+
+## 1.0.10 (June 6, 2022)
 
 ENHANCEMENT:
 * Added support for optional `incident_urgency_rule`, `support_hours`, & `scheduled_actions`.
@@ -6,27 +28,27 @@ ENHANCEMENT:
 CHORE:
 * Added outputs for PagerDuty URL, Id, and SNS Topic ARN.
 
-## pagerduty-v1.0.9 (June 3, 2022)
+## 1.0.9 (June 3, 2022)
 
 CHORE:
 * Updated PagerDuty Service provider to v2.5.0
 
-## pagerduty-v1.0.8 (May 27, 2022)
+## 1.0.8 (May 27, 2022)
 
 CHORE:
 * Added description property to the PagerDuty service.
 
-## pagerduty-v1.0.7 (May 27, 2022)
+## 1.0.7 (May 27, 2022)
 
 CHORE:
 * Added required field values in readme.
 
-## pagerduty-v1.0.6 (May 27, 2022)
+## 1.0.6 (May 27, 2022)
 
 CHORE:
 * Added default values back to readme.
 
-## pagerduty-v1.0.5 (May 27, 2022)
+## 1.0.5 (May 27, 2022)
 
 BUG:
 * PagerDuty integration key was not found (typo)
@@ -41,27 +63,27 @@ CHORE:
 EHANCEMENT:
 * Made schema webhook a variable to accommodate all webhook types (Slack V2, Generic Webhook, Generic Webhook V1, etc.)
 
-## pagerduty-v1.0.4 (February 24, 2022)
+## 1.0.4 (February 24, 2022)
 
 CHORE:
 * Added descriptions to outputs.
-## pagerduty-v1.0.3 (February 24, 2022)
+## 1.0.3 (February 24, 2022)
 
 CHORE:
 * Corrected typo for pd_subscription
 
-## pagerduty-v1.0.2 (February 24, 2022)
+## 1.0.2 (February 24, 2022)
 
 CHORE:
 * Set resource attributes to best practices
 
-## pagerduty-v1.0.1 (February 24, 2022)
+## 1.0.1 (February 24, 2022)
 
 * Updated README
 * Added CHANGELOG
 * Cleaned up variables
 
-## pagerduty-v1.0.0 (February 21, 2022)
+## 1.0.0 (February 21, 2022)
 
 INITIAL: